Add max length to spawned cut ropes
Previously, if the line trace does not detect where the rope is connected, then the rope will be set to spawn with a length equal to the trace length and with no end point.

Now, if the rope is set to spawn with a length over a default value (800), then no rope is spawned.